Update timestamp and add data.db to .gitignore
