+++ .gitignore
@@ -153,4 +153,4 @@ pnpm-lock.yaml
 
 dist
 
-
+data.db
