+++ package.json
@@ -65,6 +65,7 @@
     "openai": "^4.14.2",
     "pm2": "^5.3.0",
     "request": "^2.88.2",
-    "require": "^2.4.20"
+    "require": "^2.4.20",
+    "sqlite3": "^5.1.7"
   }
 }
